refractor nms modules to allow for no javadocing, publishes javadocs to repo(s)
