--- pineapple-nms/api/src/main/java/sh/miles/nms/api/loader/NMSManager.java
@@ -1,50 +1,0 @@
-package sh.miles.nms.api.loader;
-
-import org.jetbrains.annotations.ApiStatus;
-import org.jetbrains.annotations.NotNull;
-import sh.miles.nms.api.PineappleNMS;
-import sh.miles.pineapple.ReflectionUtils;
-
-import java.nio.channels.Pipe;
-
-/**
- * Provides Management for NMS Classes
- */
-public class NMSManager {
-
-    private static final String PATH = "sh.miles.pineapple.nms.%s.%s";
-    private static PineappleNMS handle = getHandle();
-
-    private NMSManager() {
-        throw new UnsupportedOperationException("That isn't very fresh of you sneaky boy ;)");
-    }
-
-    /**
-     * Gets the Pineapple
-     * <p>
-     * The pineapple(Ananas comosus) is a tropical plant with an edible fruit; it is the most economically significant
-     * plant in the family Bromeliaceae.
-     * <p>
-     * The pineapple is indigenous to South America, where it has been cultivated for many centuries. The introduction
-     * of the pineapple to Europe in the 17th century made it a significant cultural icon of luxury. Since the 1820s,
-     * pineapple has been commercially grown in greenhouses and many tropical plantations.
-     *
-     * @return the pineapple
-     */
-    @NotNull
-    public static PineappleNMS getPineapple() {
-        return handle;
-    }
-
-    /**
-     * Creates a NMS handle
-     *
-     * @return returns the type of handle
-     */
-    @ApiStatus.Internal
-    private static PineappleNMS getHandle() {
-        return ReflectionUtils.newInstance(PATH.formatted(MinecraftVersion.CURRENT.getInternalName(), Pipe.class.getSimpleName()), new Object[0]);
-    }
-
-
-}
